requirements.txt
fastapi
uvicorn[standard]
yfinance
pandas
numpy
requests
python-dotenv
